Fix: Correct types for compatibility with 64bit windows.
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@555 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CDlgCompare.cpp
@@ -62,7 +62,7 @@ int CDlgCompare::DoModal(
 	m_bIsModified = bIsModified;
 	m_pszComparePath = pszComparePath;
 	m_phwndCompareWnd = phwndCompareWnd;
-	return CDialog::DoModal( hInstance, hwndParent, IDD_COMPARE, lParam );
+	return (int)CDialog::DoModal( hInstance, hwndParent, IDD_COMPARE, lParam );
 }
 
 BOOL CDlgCompare::OnBnClicked( int wID )
